% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthmetic.R
\name{rthmetic}
\alias{rthmetic}
\title{Perform basic arithmetic operations and sanity checks}
\usage{
rthmetic(x, y, s)
}
\arguments{
\item{x}{A numeric expression.}

\item{y}{A numeric expression.}

\item{s}{An arithmetic operation sign (one of \code{"+"}, \code{"-"}, \code{"*"}, \code{"/"})}
}
\value{
A numeric expression as a result of given input.
}
\description{
Perform basic arithmetic operations and sanity checks
}
\examples{
rthmetic(2, 5, "+")
}
